refactor : 각 Entity Repository Test 생성
